+++ src/main/java/fr/alexandreklotz/quickdesk/controller/PermissionController.java
@@ -1,0 +1,4 @@
+package fr.alexandreklotz.quickdesk.controller;
+
+public class PermissionController {
+}