+++ src/main/java/fr/alexandreklotz/quickdesk/controller/RoleController.java
@@ -1,0 +1,12 @@
+package fr.alexandreklotz.quickdesk.controller;
+
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@CrossOrigin
+public class RoleController {
+
+
+
+}